requirements.txt
Django
django-suit
django-suit-locale
django-import-export
jira-python
